% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldie.R
\docType{package}
\name{oldie-package}
\alias{oldie}
\alias{oldie-package}
\title{oldie: Tools for Deprecating and Retiring Functions or Arguments}
\description{
A toolbox for marking functions or arguments as
deprecated with different levels of verbosity defined by a release
cycle.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/oldie}
  \item Report bugs at \url{https://github.com/r-lib/oldie/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
