% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-retired.R
\name{promote_retirement}
\alias{promote_retirement}
\alias{demote_retirement}
\title{Promote or demote retirement levels}
\usage{
promote_retirement()

demote_retirement()
}
\description{
When retirement levels are promoted, soft-deprecated functions
issue a warning and deprecated functions issue an error. There is
no change for defunct functions.
}
\details{
You can check whether deprecation levels are promoted by inspecting
the global option \code{oldie_verbose_retirement}.
}
