% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-deprecated.R
\name{signal_deprecated}
\alias{signal_deprecated}
\title{Signal deprecated function or arguments}
\usage{
signal_deprecated(.fn, .cycle, ..., .msg = NULL)
}
\arguments{
\item{.fn}{The function to deprecate or whose arguments are to be
deprecated. This should be supplied as a bare name.}

\item{.cycle}{A character vector defining the deprecation cycle.}

\item{...}{Replacements.}

\item{.msg}{A custom error message.}
}
\description{
Signal deprecated function or arguments
}
