% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-deprecation.R
\name{signal_deprecation}
\alias{signal_deprecation}
\title{Signal deprecation of a function or arguments}
\usage{
signal_deprecation(.fn, .cycle, ..., .msg = NULL)
}
\arguments{
\item{.fn}{The function to deprecate or whose arguments are to be
deprecated. This should be supplied as a bare name.}

\item{.cycle}{A character vector defining the deprecation cycle.}

\item{...}{Replacements.}

\item{.msg}{A custom error message.}
}
\description{
Signal deprecation of a function or arguments
}
